MVP Pagination
Users query was paginated.
Events was not as I hope we do not have more than 100 events in a week.
Event details is a 1 record GET.
The Emails send is a POST so no pagination there.

I also moved some of the variables to class level globals so that there is not so much passing, this will help when I start some refactor later.